Consolidate storage services
