+++ teachertool/src/transforms/setAutorun.ts
@@ -1,11 +1,11 @@
 import { stateAndDispatch } from "../state";
 import * as Actions from "../state/actions";
-import * as LocalStorage from "../services/localStorage";
+import * as Storage from "../services/storageService";
 import * as AutorunService from "../services/autorunService";
 
 export function setAutorun(autorun: boolean) {
     const { dispatch } = stateAndDispatch();
     dispatch(Actions.setAutorun(autorun));
-    LocalStorage.setAutorun(autorun);
+    Storage.setAutorun(autorun);
     AutorunService.poke();
 }
